backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
starlette==0.27.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-dotenv
httpx
opencv-python
pillow
numpy==1.24.3
openai==0.28.0
transformers==4.33.1
pymongo==4.5.0
motor==3.3.1
sqlalchemy==2.0.21
asyncpg==0.28.0 